replace i3term with "terminal_command"
this is a terminal (and therefore wm) agnostic way of starting a terminal.
we also create a terminal_lock file which is also removed by a rm trap on EXIT
in "file_cmd" (the script executed in the new terminal). this is because
some terminals commands fork to bg, others dont..